fix typing

+++ Core/Generators/BotLevelGenerator.cs
@@ -42,7 +42,7 @@ public class BotLevelGenerator
         var botLevelRange = GetRelativeBotLevelRange(botGenerationDetails, levelDetails, expTable.Length);
 
         // Get random level based on the exp table.
-        double exp = 0;
+        int exp = 0;
         var level = int.Parse(ChooseBotLevel(botLevelRange.Min.Value, botLevelRange.Max.Value, 1, 1.15).ToString()); // TODO - nasty double to string to int conversion
         for (var i = 0; i < level; i++)
         {
@@ -52,7 +52,7 @@ public class BotLevelGenerator
         // Sprinkle in some random exp within the level, unless we are at max level.
         if (level < expTable.Length - 1)
         {
-            exp += _randomUtil.GetDouble(0, expTable[level].Experience.Value - 1);
+            exp += _randomUtil.GetInt(0, expTable[level].Experience.Value - 1);
         }
 
         return new RandomisedBotLevelResult{ Level = level, Exp = exp };
